Bias towards write efficiency
The original implementation biased towared keeping a Histogram's internal data sorted on mutation, which allowed for trivial percentile lookups. Instead, track a dirty flag, and sort when getting a percentile if necessary.

This is no longer strictly thread-safe. To be so, the observe and percentile code (specifically for sorting) would ned to be made exclusive, which is unlikely an a problem in practice, but the overhead for locking would remove most of the gains we just realized.